backend/pyproject.toml
[project]
name = "reliant-scheduler"
version = "0.1.0"
description = "Enterprise workload automation and job scheduling platform"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115,<1",
    "uvicorn[standard]>=0.34,<1",
    "sqlalchemy>=2.0,<3",
    "alembic>=1.14,<2",
    "asyncpg>=0.30,<1",
    "pydantic>=2.10,<3",
    "pydantic-settings>=2.7,<3",
    "azure-servicebus>=7.13,<8",
    "azure-eventhub>=5.14,<6",
    "azure-storage-blob>=12.24,<13",
    "azure-identity>=1.19,<2",
    "azure-keyvault-secrets>=4.9,<5",
    "azure-monitor-opentelemetry>=1.6,<2",
    "httpx>=0.28,<1",
    "structlog>=24.4,<25",
    "apscheduler>=3.11,<4",
    "croniter>=5.0,<6",
    "prometheus-client>=0.21,<1",
    "PyJWT[crypto]>=2.9,<3",
    "msal>=1.31,<2",
    "asyncssh>=2.17,<3",
    "pypsrp>=0.8,<1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3,<9",
    "pytest-asyncio>=0.25,<1",
    "pytest-cov>=6.0,<7",
    "httpx>=0.28,<1",
    "ruff>=0.9,<1",
    "mypy>=1.14,<2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = ["azure: Tests requiring Azure services or emulators"]

[tool.coverage.run]
source = ["reliant_scheduler"]
omit = ["*/tests/*"]

[tool.coverage.report]
fail_under = 70
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.",
]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]

[tool.hatch.build.targets.wheel]
packages = ["src/reliant_scheduler"]
